Debian stretch: add complete Debian repositories
To get a recent Debian image, distro "stretch" is not sufficient, but we need
stretch-updates plus stretch/updates from security.debian.org.

This patch is a WIP - it lacks corresponding changes to multistrap.conf.in from
recipes-devtools/buildchroot and the image recipes for Raspbian (and
older Debian versions?).

Based on an initial patch by Henning Schild <henning.schild@siemens.com>.

Signed-off-by: Gernot Hillier <gernot.hillier@siemens.com>