meta/base.bbclass: integrate COMPATIBLE_MACHINE feature
Integrate the COMPATIBLE_MACHINE of OE, where recipes can specify a
regex in COMPATIBLE_MACHINE that needs to match to an entry of the
MACHINEOVERRIDES variable, otherwise the recipe will be skipped.

Similar to OE, with `PARSE_ALL_RECIPES`, parsing of the recipe can be
enforced.

This code was copied from the OE-core codebase.

Signed-off-by: Claudius Heine <ch@denx.de>