beagleplay: Add U-Boot recipe
Upstream support is close but not there yet. Base the support on
upcoming 2023.10, currently with a few extra patches.

We customize the boot by preferring the SD card over the eMMC and
skipping all the TI specialties which may pull the eMMC to the front
again, directly going for distro boot.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>